planCharge-41 Ajouter un module "Disponibilités" - Dév/évo/fonct : La sauvegarde inclut maintenant les pctages de dispo pour la CT. Pas testé le chargement. ET la table reste tjs vide alors que les données sont bien en mémoire.

+++ planCharge-ihm/src/main/java/fr/gouv/agriculture/dal/ct/planCharge/ihm/model/disponibilite/NbrsJoursAbsenceBean.java
@@ -3,14 +3,14 @@ package fr.gouv.agriculture.dal.ct.planCharge.ihm.model.disponibilite;
 import fr.gouv.agriculture.dal.ct.ihm.model.BeanException;
 import fr.gouv.agriculture.dal.ct.metier.dto.AbstractDTO;
 import fr.gouv.agriculture.dal.ct.planCharge.ihm.model.referentiels.RessourceHumaineBean;
-import javafx.beans.property.DoubleProperty;
+import javafx.beans.property.FloatProperty;
 
 import javax.validation.constraints.NotNull;
 import java.time.LocalDate;
 import java.util.Map;
 import java.util.TreeMap;
 
-public class NbrsJoursAbsenceBean extends AbstractDisponibilitesRessourceBean<AbstractDTO, NbrsJoursAbsenceBean, DoubleProperty> {
+public class NbrsJoursAbsenceBean extends AbstractDisponibilitesRessourceBean<AbstractDTO, NbrsJoursAbsenceBean, FloatProperty> {
 
 
     // Fields:
@@ -21,7 +21,7 @@ public class NbrsJoursAbsenceBean extends AbstractDisponibilitesRessourceBean<Ab
         super(ressourceHumaineBean, new TreeMap<>()); // TreeMap juste pour trier afin de faciliter le débogage.
     }
 
-    public NbrsJoursAbsenceBean(@NotNull RessourceHumaineBean ressourceHumaineBean, @NotNull Map<LocalDate, DoubleProperty> calendrier) {
+    public NbrsJoursAbsenceBean(@NotNull RessourceHumaineBean ressourceHumaineBean, @NotNull Map<LocalDate, FloatProperty> calendrier) {
         super(ressourceHumaineBean, calendrier);
     }
